Refactored everything as services on Dictionary
Now everything is super clean interface-wise, which allows
me to make functions on the dictionary that have interplay with each
other, whilst keeping the code clean, e.g. converting word information
with lookups and whatnot.

Will comment more on actual functions later. The behaviour is the same but this sets up for future cleanliness and testing and whatnot.