Finally have decent word-wrap 😤
i realised there were several issues with the existing word-wrapping:

it was putting a space directly after a newline, and was cutting off
unneccessarily. i found that this was because internal newlines (e.g.
newlines that were already present) were not being split, as i was only
splitting on " " of course. I had to spend the time to figure out a
lousy O(n*m) algorithm (maybe it's even worse idk) to split on newlines,
because at the beginning of the word wrap, ALL newlines that were in
the string should remain there.

This skipping of newlines as part of words was throwing off the
space/length calculator mucho.

e.g. "this\n1.list\n2.of things" should tokenise to [this, \n, 1.list,
\n, 2.of things] and then depending on how much space is on the line,
get more newlines inserted to perform wrapping.

Whew!