stupid refactor
intended to split things off into seperate packages in the ui/ dir but
that kept giving me import cycles. anyway, i realised that having
everything in Layout() is wrong, so i removed the gocui dep from lis.go
and put a gocui.Gui object on the ViewManager, then i made a run
function which instantiates the views and makes the runloop. this is
much better